+++ src/main/java/uk/q3c/krail/core/i18n/DefaultI18NFieldScanner.java
@@ -11,15 +11,17 @@
  *
  */
 
-package uk.q3c.krail.i18n;
+package uk.q3c.krail.core.i18n;
 
 import com.google.inject.Inject;
 import com.vaadin.ui.AbstractComponent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import uk.q3c.krail.i18n.I18NFieldScanner;
 import uk.q3c.krail.i18n.api.I18N;
 import uk.q3c.krail.i18n.api.I18NAnnotation;
 import uk.q3c.krail.i18n.api.I18NException;
+import uk.q3c.util.clazz.UnenhancedClassIdentifier;
 
 import javax.annotation.Nonnull;
 import java.lang.annotation.Annotation;
@@ -37,11 +39,11 @@ public class DefaultI18NFieldScanner implements I18NFieldScanner {
     private static Logger log = LoggerFactory.getLogger(DefaultI18NFieldScanner.class);
     private Map<AbstractComponent, AnnotationInfo> components;
     private LinkedList<Object> drillDowns;
-    private I18NHostClassIdentifier i18NHostClassIdentifier;
+    private UnenhancedClassIdentifier i18NHostClassIdentifier;
     private List<Object> processedDrillDowns;
 
     @Inject
-    protected DefaultI18NFieldScanner(I18NHostClassIdentifier i18NHostClassIdentifier) {
+    protected DefaultI18NFieldScanner(UnenhancedClassIdentifier i18NHostClassIdentifier) {
         this.i18NHostClassIdentifier = i18NHostClassIdentifier;
         components = new HashMap<>();
         drillDowns = new LinkedList<>();