+++ src/main/java/uk/q3c/krail/i18n/DefaultI18NProcessor.java
@@ -21,6 +21,7 @@ import com.vaadin.ui.Table;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import uk.q3c.krail.core.i18n.AnnotationInfo;
 import uk.q3c.krail.core.ui.ScopedUI;
 import uk.q3c.krail.i18n.api.CurrentLocale;
 import uk.q3c.krail.i18n.api.I18NException;