+++ src/main/java/uk/q3c/krail/i18n/VaadinI18NModule.java
@@ -19,6 +19,7 @@ import com.google.inject.multibindings.MapBinder;
 import com.google.inject.multibindings.Multibinder;
 import uk.q3c.krail.core.guice.uiscope.UIScoped;
 import uk.q3c.krail.core.guice.vsscope.VaadinSessionScoped;
+import uk.q3c.krail.core.i18n.DefaultI18NFieldScanner;
 import uk.q3c.krail.core.i18n.LabelKey;
 import uk.q3c.krail.core.option.Option;
 import uk.q3c.krail.core.persist.cache.i18n.DefaultPatternCacheLoader;
@@ -27,6 +28,8 @@ import uk.q3c.krail.core.persist.common.common.KrailPersistenceUnitHelper;
 import uk.q3c.krail.i18n.api.*;
 import uk.q3c.krail.i18n.api.clazz.ClassPatternDao;
 import uk.q3c.krail.i18n.api.clazz.ClassPatternSource;
+import uk.q3c.util.clazz.DefaultUnenhancedClassIdentifier;
+import uk.q3c.util.clazz.UnenhancedClassIdentifier;
 
 import javax.annotation.Nonnull;
 import java.lang.annotation.Annotation;
@@ -150,10 +153,10 @@ public class VaadinI18NModule extends I18NModule {
 
 
     /**
-     * See javadoc for {@link I18NHostClassIdentifier} for an explanation of what this is for.  Override this method if you provide your own implementation
+     * See javadoc for {@link UnenhancedClassIdentifier} for an explanation of what this is for.  Override this method if you provide your own implementation
      */
     protected void bindHostClassIdentifier() {
-        bind(I18NHostClassIdentifier.class).to(DefaultI18NHostClassIdentifier.class);
+        bind(UnenhancedClassIdentifier.class).to(DefaultUnenhancedClassIdentifier.class);
     }
 
     /**