+++ src/main/java/uk/q3c/krail/i18n/api/DefaultActivePatternDao.java
@@ -11,10 +11,9 @@
  *
  */
 
-package uk.q3c.krail.i18n;
+package uk.q3c.krail.i18n.api;
 
 import com.google.inject.BindingAnnotation;
-import uk.q3c.krail.i18n.api.PatternDao;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -23,7 +22,7 @@ import static java.lang.annotation.ElementType.*;
 import static java.lang.annotation.RetentionPolicy.*;
 
 /**
- * Binding annotation to identify the active {@link PatternDao} (where active is the one used by Krail core)
+ * Binding annotation to identify the active {@link PatternDao}
  * <p>
  * <p>
  * Created by David Sowerby on 06/02/15.