+++ src/test/java/uk/q3c/krail/i18n/I18NTestClass.java
@@ -16,6 +16,8 @@ import com.vaadin.ui.Button;
 import com.vaadin.ui.Grid;
 import com.vaadin.ui.Label;
 import com.vaadin.ui.Table;
+import uk.q3c.krail.core.i18n.Caption;
+import uk.q3c.krail.core.i18n.Description;
 import uk.q3c.krail.core.i18n.DescriptionKey;
 import uk.q3c.krail.core.i18n.LabelKey;
 import uk.q3c.krail.i18n.api.I18N;