+++ src/test/java/uk/q3c/krail/i18n/I18NTestClass7.java
@@ -14,6 +14,7 @@
 package uk.q3c.krail.i18n;
 
 
+import uk.q3c.krail.core.i18n.Caption;
 import uk.q3c.krail.core.i18n.DescriptionKey;
 import uk.q3c.krail.core.i18n.LabelKey;